Support using system go in /usr/local/go

+++ main.go
@@ -46,25 +46,29 @@ func main() {
 	// Location on the filesystem to store the golang archive
 	golangArchive := filepath.Join(baseDir, path.Base(fileUrl))
 
-	err = downloadFile(fileUrl, golangArchive)
-	if err != nil {
-		log.Fatalf("Failed to download: %v", err)
-	}
-
-	// Extract golang archive
-	canaryFile := filepath.Join(baseDir, EXTRACTED_CANARY) // File that signals extraction has already occurred
-	if fileExists(canaryFile) {
-		log.Debugf("Skipping extraction due to presence of canary at %q", canaryFile)
-	} else {
-		// Remove extracted canary, if exists
-		_ = os.Remove(filepath.Join(baseDir, EXTRACTED_CANARY))
-
-		err = extractFile(golangArchive, baseDir)
+	if version != "system" {
+		err = downloadFile(fileUrl, golangArchive)
 		if err != nil {
-			log.Fatalf("Failed to extract: %v", err)
+			log.Fatalf("Failed to download: %v", err)
 		}
-		ioutil.WriteFile(canaryFile, []byte(""), 0755)
-		log.Infof("Successfully extracted %q", golangArchive)
+
+		// Extract golang archive
+		canaryFile := filepath.Join(baseDir, EXTRACTED_CANARY) // File that signals extraction has already occurred
+		if fileExists(canaryFile) {
+			log.Debugf("Skipping extraction due to presence of canary at %q", canaryFile)
+		} else {
+			// Remove extracted canary, if exists
+			_ = os.Remove(filepath.Join(baseDir, EXTRACTED_CANARY))
+
+			err = extractFile(golangArchive, baseDir)
+			if err != nil {
+				log.Fatalf("Failed to extract: %v", err)
+			}
+			ioutil.WriteFile(canaryFile, []byte(""), 0755)
+			log.Infof("Successfully extracted %q", golangArchive)
+		}
+	} else {
+		baseDir = "/usr/local"
 	}
 
 	// Run go command
@@ -76,6 +80,10 @@ func main() {
 
 	err = runGo(binary, baseDir, os.Args[1:])
 	if err != nil {
-		log.Fatalf("command failed: %v", err)
+		if version == "system" {
+			log.Fatalf("Failed to run system go: %v", err)
+		} else {
+			log.Fatalf("command failed: %v", err)
+		}
 	}
 }
